requirements.txt
Pillow
PySimpleGUI
